requirements.txt
opencv-python
openpyxl
Pillow